Cleaning things up :)
